requirements.txt
PyPDF2>=3.0.1
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
streamlit>=1.28.0
requests>=2.28.0requests>=2.28.0
